+++ lib/vanagon/platform/defaults/ubuntu-18.04-aarch64.rb
@@ -4,7 +4,24 @@ platform "ubuntu-18.04-aarch64" do |plat|
   plat.servicetype "systemd"
   plat.codename "bionic"
 
-  packages = %w(build-essential devscripts make quilt pkg-config debhelper rsync fakeroot cmake)
+  packages = %w(
+    autoconf
+    build-essential
+    cmake
+    debhelper
+    devscripts
+    fakeroot
+    libbz2-dev
+    libreadline-dev
+    libselinux1-dev
+    openjdk-8-jre-headless
+    pkg-config
+    quilt
+    rsync
+    swig
+    systemtap-sdt-dev
+    zlib1g-dev
+  )
   plat.provision_with "export DEBIAN_FRONTEND=noninteractive; apt-get update -qq; apt-get install -qy --no-install-recommends #{packages.join(' ')}"
   plat.install_build_dependencies_with "DEBIAN_FRONTEND=noninteractive; apt-get install -qy --no-install-recommends "
   plat.vmpooler_template "ubuntu-1804-arm64"